feat(EntityCharges): Add EntityChargesPage component for managing charges on addons and plans, implement bulk price creation API, and update routing for addon charges
