+++ src/components/organisms/EntityChargesPage/index.ts
@@ -1,0 +1,2 @@
+export { default } from './EntityChargesPage';
+export { ENTITY_TYPE } from './EntityChargesPage';
