+++ src/components/organisms/index.ts
@@ -1,4 +1,5 @@
 export { default as EmptyPage } from './EmptyPage';
+export { default as EntityChargesPage } from './EntityChargesPage';
 export {
 	AddonTable,
 	Preview,
