+++ src/pages/product-catalog/addons/AddonCharges.tsx
@@ -1,0 +1,10 @@
+import { useParams } from 'react-router-dom';
+import EntityChargesPage, { ENTITY_TYPE } from '@/components/organisms/EntityChargesPage';
+
+const AddonChargesPage = () => {
+	const { addonId } = useParams<{ addonId: string }>();
+
+	return <EntityChargesPage entityType={ENTITY_TYPE.ADDON} entityId={addonId!} entityName='Addon' />;
+};
+
+export default AddonChargesPage;
